% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraphCssImport.R
\name{ggiraphCss}
\alias{ggiraphCss}
\title{Convenience function to embed/bake ggiraph css directly into a created HTML file}
\usage{
ggiraphCss()
}
\description{
* For ggiraph graphics: if not imported, saveHTML might to proprely import ggiraph css leading to responsiveness issues with IE
}
\examples{
\dontrun{
  ggiraphCss()
  normalizeCss()
  bootstrapCss()
}
}
