% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionWatcher.R
\docType{data}
\name{dimensionWatcher}
\alias{dimensionWatcher}
\title{html code to call Tamedia dimensionWatcher}
\format{An object of class \code{character} of length 1.}
\usage{
dimensionWatcher
}
\description{
snippet to include in page of dynamic sizing of iframe
}
\examples{
\dontrun{
dimensionWatcher
}
}
\keyword{datasets}
