% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlGrid.R
\name{dq_grid}
\alias{dq_grid}
\title{HTML responsive grid for static charts}
\usage{
dq_grid(
  ...,
  ncol = NULL,
  rowheight = NULL,
  add_htmlgrid_css = TRUE,
  browsable = TRUE
)
}
\arguments{
\item{...}{either individual `ggplot2` objects or a `list` of `ggplot2` objects.}

\item{ncol}{how many columns in the grid}

\item{rowheight}{Height in px.}

\item{add_htmlgrid_css}{A logical value to add or not `htmlgrid.css` as
dependency.}

\item{browsable}{Logical value indicating if the returned object is converted
to an HTML object browsable using \code{htmltools::browsable}.}
}
\description{
Lays out ggplot2 charts into a "grid" (similar to
\code{grid.arrange} from \code{gridExtra}) but in 
a responsive grid, html/css output.
}
\examples{
charts <- lapply(1:9, function(x) {
  qplot(cumsum(rnorm(100)), geom = "histogram")  + theme_tam()
})
if(interactive()){
  dq_grid(charts, rowheight = 300)
}
}
