% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_update_chart_ifnewdata.R
\name{dw_update_chart_ifnewdata}
\alias{dw_update_chart_ifnewdata}
\title{Update datawrapper chart if new data is available}
\usage{
dw_update_chart_ifnewdata(
  dw_id,
  api_key,
  new_data,
  annotate_text = "",
  timestamp_update,
  timestamp_text = "Actualisé le ",
  date_format = "\%d \%B \%Y"
)
}
\arguments{
\item{dw_id}{a character sclalar, datawrapper id}

\item{api_key}{a character scalar, datawrapper api key}

\item{new_data}{a data.frame, the new data}

\item{annotate_text}{a character scalar, the text for the annotation that will be before the timestamp}

\item{timestamp_update}{a logical scalar, if TRUE, the timestamp will be added at bottom of the chart}

\item{timestamp_text}{a character scalar, the text for the timestamp}

\item{date_format}{a character scalar, the date format}
}
\value{
a boolean scalar. TRUE if the chart is updated, FALSE otherwise
}
\description{
Update datawrapper chart if new data is available
}
