% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_theme_temps.R
\docType{data}
\name{hc_theme_temps}
\alias{hc_theme_temps}
\title{Nouveau thème Highcharter Le Temps basé sur le styleguide 2024}
\format{
An object of class \code{hc_theme} of length 11.
}
\usage{
hc_theme_temps
}
\description{
Thème moderne pour highcharter suivant les nouvelles spécifications visuelles de Le Temps
}
\examples{
\dontrun{
require(highcharter)
require(dqnTheme)
data(diamonds, mpg, package = "ggplot2")

hchart(mpg, "scatter", 
  hcaes(x = displ, y = hwy, group = class)) \%>\% 
  hc_add_theme(hc_theme_temps) \%>\% 
  hc_title(text = "Titre principal \u2013 Max 40 caract\u00e8res") \%>\% 
  hc_subtitle(text = "Description \u2013 Max 150 caract\u00e8res. Aliquam dictum dolor tellus, vel egestas lacus mollis.")
}
}
\seealso{
http://jkunst.com/highcharter/themes.html#create-themes
}
\keyword{datasets}
