% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter_theme.R
\docType{data}
\name{hc_theme_tg}
\alias{hc_theme_tg}
\alias{hc_theme_TA}
\alias{style_tam_highcharter}
\alias{style_ta_highcharter}
\alias{htmlwidgetgrid_css}
\title{Highchart(er) TG-Tamedia theme}
\format{
An object of class \code{hc_theme} of length 8.

An object of class \code{hc_theme} of length 8.
}
\usage{
hc_theme_tg

hc_theme_TA

style_tam_highcharter()

style_ta_highcharter()

htmlwidgetgrid_css()
}
\description{
Some helpers to get a TG look for highcharter
}
\examples{
require(highcharter)
require(dqnTheme)
data(diamonds, mpg, package = "ggplot2")
hchart(mpg, "scatter", 
  hcaes(x = displ, y = hwy, group = class)) \%>\% 
  hc_add_theme(hc_theme_tg) \%>\% 
  hc_title(text = "super catchy title") \%>\% 
  hc_subtitle(text = "a description of what the hell is shown")
require(highcharter)
require(dqnTheme)
data(diamonds, mpg, package = "ggplot2")
hchart(mpg, "scatter", 
  hcaes(x = displ, y = hwy, group = class)) \%>\% 
  hc_add_theme(hc_theme_TA) \%>\% 
  hc_title(text = "Super catchy title") \%>\% 
  hc_subtitle(text = "a description of what the hell is shown")
style_tam_highcharter()
style_ta_highcharter()
# A modified version of highcharter's htmlwdgtgrid.css, with much less margin on mobile
htmlwidgetgrid_css()
}
\seealso{
http://jkunst.com/highcharter/themes.html#create-themes
}
\keyword{datasets}
