% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_temps.R, R/temps_theme.R
\docType{data}
\name{temps_pal}
\alias{temps_pal}
\alias{temps_partis_pal}
\alias{temps_seq1_pal}
\alias{temps_seq2_pal}
\alias{temps_seq3_pal}
\alias{temps_div1_pal}
\alias{temps_div2_pal}
\alias{temps_div3_pal}
\alias{get_temps_sequential}
\alias{get_temps_divergent}
\alias{get_temps_partis}
\alias{setup_temps_fonts}
\title{Palettes de couleurs Le Temps}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 5.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 9.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 5.
}
\usage{
temps_pal

temps_partis_pal

temps_seq1_pal

temps_seq2_pal

temps_seq3_pal

temps_div1_pal

temps_div2_pal

temps_div3_pal

get_temps_sequential(type = 1, n = 7)

get_temps_divergent(type = 1, n = 5)

get_temps_partis(partis)

temps_pal

temps_partis_pal

temps_seq1_pal

temps_div1_pal

setup_temps_fonts()
}
\arguments{
\item{type}{type de palette divergente (1, 2, ou 3)}

\item{n}{nombre de couleurs désirées}

\item{partis}{vecteur des noms de partis}
}
\description{
Toutes les palettes de couleurs selon le styleguide Le Temps 2024
}
\examples{
# Palette principale
pie(rep(1, length(temps_pal)), col = temps_pal)

# Couleurs des partis
barplot(1:7, col = temps_partis_pal)
\dontrun{
# Configurer les polices avant d'utiliser le thème
setup_temps_fonts()

# Puis utiliser le thème normalement
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  theme_temps()
}
}
\keyword{datasets}
