% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamTheme.R
\name{theme_tam}
\alias{theme_tam}
\title{Tribune de Genève / Tamedia Romandie chart theme}
\usage{
theme_tam(ticks = TRUE, base_family = "Titillium Web", base_size = 13)
}
\arguments{
\item{ticks}{\code{logical} Show axis ticks?}

\item{base_family}{Base font family}

\item{base_size}{Base font size}
}
\description{
ggplot  theme copied/inspired from https://gist.github.com/hrbrmstr/283850725519e502e70c
}
\examples{
require(ggplot2)
require(tamTheme)
qplot(1:10, 1:10, size = 10:1) + xlab("axis x label") + ylab ("y axis label") + theme_tam()
qplot(mtcars$mpg) + theme_tam()


}
