% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{tam_pal}
\alias{tam_pal}
\alias{tam_dpal}
\alias{tam_spal}
\title{Tamedia TG color palette}
\format{An object of class \code{character} of length 8.}
\usage{
tam_pal

tam_dpal

tam_spal
}
\description{
Just a bunch of hex color vectors... 
Color palette as vector (no function). Colors can be sequential (one color), diverging (2 colors)
}
\details{
tam_pal: 8 categorical tam colors

tam_dpal: diverging color palette with 13 levels

tam_spal:  single hue blueish
}
\examples{
pie(rep(1,length(tam_pal)), col=tam_pal)
pie(rep(1,length(tam_dpal)), col=tam_dpal)
pie(rep(1,length(tam_spal)), col=tam_spal)
}
\keyword{datasets}
