% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_button_helper.R
\name{dw_button_helper}
\alias{dw_button_helper}
\title{Generate Responsive HTML buttons for Datawrapper}
\usage{
dw_button_helper(
  dw_id,
  dw_label,
  bg_col_hl = "#414241",
  bg_col_norm = "#BFC0BF",
  font_col_hl = "#FFFFFF",
  font_col_norm = "#FFFFFF",
  font_weight_hl = 600,
  font_weight_norm = 400
)
}
\arguments{
\item{dw_id}{A character vector of Datawrapper IDs.}

\item{dw_label}{A character vector of labels for the buttons.}

\item{bg_col_hl}{Background color of the highlighted button.}

\item{bg_col_norm}{Background color of the normal button.}

\item{font_col_hl}{Font color of the highlighted button.}

\item{font_col_norm}{Font color of the normal button.}

\item{font_weight_hl}{Font weight of the highlighted button.}

\item{font_weight_norm}{Font weight of the normal button.}
}
\value{
A character vector of HTML strings.
}
\description{
Create clickable, responsive buttons to navigate different Datawrapper charts
}
