% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_button_helper.R
\name{dw_button_helper}
\alias{dw_button_helper}
\title{Generate HTML buttons for datawrapper}
\usage{
dw_button_helper(
  dw_id,
  dw_label,
  bg_col_hl = "#414241",
  bg_col_norm = "#BFC0BF",
  font_col_hl = "#FFFFFF",
  font_col_norm = "#FFFFFF",
  font_weight_hl = 800,
  font_weight_norm = 400
)
}
\arguments{
\item{dw_id}{a character vector of datawrapper ids}

\item{dw_label}{a character vector of labels for the buttons}

\item{bg_col_hl}{a character string for the background color of the highlighted button}

\item{bg_col_norm}{a character string for the background color of the normal button}

\item{font_col_hl}{a character string for the font color of the highlighted button}

\item{font_col_norm}{a character string for the font color of the normal button}

\item{font_weight_hl}{a numeric value for the font weight of the highlighted button}

\item{font_weight_norm}{a numeric value for the font weight of the normal button}
}
\value{
a character string
}
\description{
Create clickable buttons to navigate different datawrapper charts
  
dw_button_helper
}
\examples{
dw_id <- c("tXf1C", "6MxFK")
dw_label <- c("2025", "2021")
dw_button_helper(dw_id, dw_label) 
}
