% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dw_update_chart_ifnewdata.R
\name{fuzzy_compare}
\alias{fuzzy_compare}
\alias{dw_update_chart_ifnewdata}
\title{Update datawrapper chart if new data is available}
\usage{
fuzzy_compare(df1, df2, tolerance = 0.1, ignore_case = TRUE)

dw_update_chart_ifnewdata(
  dw_id,
  api_key,
  new_data,
  annotate_text = "",
  timestamp_update = F,
  timestamp_text = "Actualisé le ",
  date_format = "\%d \%B \%Y"
)
}
\arguments{
\item{df1, df2}{data frames to compare}

\item{tolerance}{a numeric scalar, the tolerance for comparison}

\item{ignore_case}{a logical scalar, if TRUE, the comparison will be case-insensitive}

\item{dw_id}{a character sclalar, datawrapper id}

\item{api_key}{a character scalar, datawrapper api key}

\item{new_data}{a data.frame, the new data}

\item{annotate_text}{a character scalar, the text for the annotation that will be before the timestamp}

\item{timestamp_update}{a logical scalar, if TRUE, the timestamp will be added at bottom of the chart}

\item{timestamp_text}{a character scalar, the text for the timestamp}

\item{date_format}{a character scalar, the date format}
}
\value{
a boolean scalar. TRUE if similar, FALSE otherwise

a boolean scalar. TRUE if the chart is updated, FALSE otherwise
}
\description{
Update datawrapper chart if new data is available
}
