% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatBreaks.R
\name{formatBreaks}
\alias{formatBreaks}
\title{Helper for more readable labels when use with \code{cut}}
\usage{
formatBreaks(brks, inbetween_sep = "-", prefix_first = "< ",
  prefix_last = "> ", suffix_first = "", suffix_last = "")
}
\arguments{
\item{brks}{numeric vector that will be used to \code{cut}}

\item{inbetween_sep, prefix_first, prefix_last, suffix_first, suffix_last}{strings that are put in between intervals, at the beginning or at the intervals only.}
}
\value{
a character string
}
\description{
Helper for more readable labels when use with \code{cut}
}
\examples{
cont_var <- rnorm (100) * 100
brks <- pretty(cont_var, n = 10) # or range(cont_var) \%>\% pretty()
brks_labels <- formatBr
eaks(brks, inbetween_sep = ' - ')
formatBreaks(brks, inbetween_sep = ' - ', suffix_first = " unité")
cut(cont_var, breaks = brks, labels = brks_labels)
}
